+++ src/commandHandlers/stop_recording.ts
