+++ src/commandHandlers/stop_recording_all.ts
